+++ lib/ohai/plugins/passwd.rb
@@ -12,10 +12,6 @@ Ohai.plugin(:Passwd) do
     str
   end
 
-  def powershell_out(ps)
-    Mixlib::ShellOut.new("powershell.exe", "-c", ps).run_command
-  end
-
   collect_data do
     require "etc" unless defined?(Etc)
 
@@ -46,42 +42,60 @@ Ohai.plugin(:Passwd) do
   end
 
   collect_data(:windows) do
+    require "wmi-lite/wmi" unless defined?(WmiLite::Wmi)
+
     unless etc
       etc Mash.new
 
+      wmi = WmiLite::Wmi.new
+
       etc[:passwd] = Mash.new
-      s = powershell_out("get-localuser | convertto-json")
-      users = JSON.parse(s.stdout)
+      users = wmi.query("SELECT * FROM Win32_UserAccount WHERE LocalAccount = True")
       users.each do |user|
         uname = user["Name"].strip.downcase
         Ohai::Log.debug("processing user #{uname}")
         etc[:passwd][uname] = Mash.new
-        user.each do |key, val|
-          etc[:passwd][uname][key.downcase] = val
+        wmi_obj = user.wmi_ole_object
+        wmi_obj.properties_.each do |key|
+          etc[:passwd][uname][key.name.downcase] = user[key.name]
         end
       end
 
       etc[:group] = Mash.new
-      s = powershell_out("get-localgroup | convertto-json")
-      groups = JSON.parse(s.stdout)
+      groups = wmi.query("SELECT * FROM Win32_Group WHERE LocalAccount = True")
       groups.each do |group|
         gname = group["Name"].strip.downcase
         Ohai::Log.debug("processing group #{gname}")
         etc[:group][gname] = Mash.new
-        group.each do |key, val|
-          etc[:group][gname][key.downcase] = val
+        wmi_obj = group.wmi_ole_object
+        wmi_obj.properties_.each do |key|
+          etc[:group][gname][key.name.downcase] = group[key.name]
         end
-        # calling this for each group is slow, but it requires
-        # a specific group, soooooo....
-        g = powershell_out(
-          "get-localgroupmember -name '#{gname}' | convertto-json"
+
+        # This is the primary reason that we're using WMI instead of powershell
+        # cmdlets - the powershell start up cost is huge, and you *must* do this
+        # query for every. single. group. individually.
+
+        # The query returns nothing unless you specify domain *and* name, it's
+        # a path, not a set of queries.
+        subq = "Win32_Group.Domain='#{group["Domain"]}',Name='#{group["Name"]}'"
+        members = wmi.query(
+          "SELECT * FROM Win32_GroupUser WHERE GroupComponent=\"#{subq}\""
         )
-        out = g.stdout
-        if !out.empty?
-          gmem = JSON.parse(g.stdout)
-          etc[:group][gname]["members"] = gmem
-        else
-          etc[:group][gname]["members"] = []
+        etc[:group][gname]["members"] = members.map do |member|
+          mi = {}
+          info = Hash[
+            member["partcomponent"].split(",").map { |x| x.split("=") }.map { |a, b| [a, b.undump] }
+          ]
+          if info.keys.any? { |x| x.match?("Win32_UserAccount") }
+            mi["type"] = :user
+          else
+            # Note: the type here is actually Win32_SystemAccount, because,
+            # that's what groups are in the Windows universe.
+            mi["type"] = :group
+          end
+          mi["name"] = info["Name"]
+          mi
         end
       end
     end